+++ packages/fighting-design/table/src/table.vue
@@ -1,12 +1,14 @@
 <script lang="ts" setup name="FTable">
   import { Props } from './props'
-  import { computed } from 'vue'
+  import { computed, ref } from 'vue'
   import { sizeChange } from '../../_utils'
   import { TableColgroupVue } from './components'
   import type { ComputedRef, CSSProperties } from 'vue'
   import type { TablePropsType } from './interface'
 
   const prop: TablePropsType = defineProps(Props)
+
+  const optionalList = ref([])
 
   /**
    * 样式列表
@@ -32,6 +34,7 @@
           <thead :align="align">
             <tr>
               <th v-if="num">序号</th>
+              <th v-if="optional">选择</th>
               <th v-for="(column, index) in columns" :key="index">
                 {{ column.title }}
               </th>
@@ -57,6 +60,9 @@
           <tbody :align="align">
             <tr v-for="(dataItem, m) in data" :key="m">
               <td v-if="num">{{ m + 1 }}</td>
+              <th v-if="optional">
+                <f-checkbox v-model="optionalList" :label="m" />
+              </th>
               <td v-for="(column, i) in columns" :key="i">
                 {{ dataItem[column.key] }}
               </td>